create separate function renderModal, move Photos interface to separate file, rename defaultAlbumsOnPage to defaultAlbumsAmount
