+++ src/modules/types.ts
@@ -1,0 +1,11 @@
+interface Photos {
+    [key: number]: {
+        albumId: number;
+        id: number;
+        thumbnailUrl: string;
+        title: string;
+        url: string;
+    }[];
+}
+
+export {Photos};
