+++ src/shiftNextImageSrc.ts
@@ -1,4 +1,4 @@
-import {imageIndex} from './renderGallery';
+import {imageIndex} from './modules/renderModal';
 //Поставить imgIndex как параметр
 function shiftNextImgSrc(imageIndex: number) {
     return imageIndex + 1;
