+++ src/shiftPrevImageSrc.ts
@@ -1,4 +1,4 @@
-import {imageIndex} from './renderGallery';
+import {imageIndex} from './modules/renderModal';
 function shiftPrevImgSrc(imageIndex: number) {
     return imageIndex - 1;
 }
